[Notifications] Mail - set subject and body of the mail
