+++ mlrun/utils/notifications/notification/mail.py
@@ -13,7 +13,8 @@
 # limitations under the License.
 import re
 import typing
-from email.message import EmailMessage
+from email.mime.multipart import MIMEMultipart
+from email.mime.text import MIMEText
 
 import aiosmtplib
 
@@ -69,9 +70,17 @@ class MailNotification(base.NotificationBase):
         alert: typing.Optional[mlrun.common.schemas.AlertConfig] = None,
         event_data: typing.Optional[mlrun.common.schemas.Event] = None,
     ):
-        message = self.params.get("message", message)
-        self.params.setdefault("subject", f"[{severity}] {message}")
-        self.params.setdefault("body", message)
+        self.params["subject"] = f"[{severity}] {message}"
+        override_body = self.params.get("override_body", None)
+
+        self.params["body"] = self._get_html(
+            message, severity, runs, custom_html, alert, event_data
+        )
+        if override_body:
+            self.params["body"] = self._serialize_runs_in_request_body(
+                override_body, runs
+            )
+
         await self._send_email(**self.params)
 
     @classmethod
@@ -146,11 +155,11 @@ class MailNotification(base.NotificationBase):
         **kwargs,
     ):
         # Create the email message
-        message = EmailMessage()
+        message = MIMEMultipart("alternative")
         message["From"] = sender_address
         message["To"] = email_addresses
         message["Subject"] = subject
-        message.set_content(body)
+        message.attach(MIMEText(body, "html"))
 
         # Send the email
         await aiosmtplib.send(
