+++ tests/utils/test_notifications.py
@@ -1027,6 +1027,8 @@ class TestMailNotification:
         "validate_certs": True,
         "start_tls": False,
     }
+    MOCKED_HTML = "mocked_html"
+    MOCKED_BODY_PLACEHOLDER = "mocked_body_placeholder"
 
     @pytest.mark.parametrize(
         "params, expectation",
@@ -1192,18 +1194,18 @@ class TestMailNotification:
                 "test-message",
                 "info",
                 {
-                    "body": "test-message",
                     "subject": "[info] test-message",
+                    "body": MOCKED_HTML,
                 },
             ),
             (
                 "with_params_message",
-                {"message": "params_message"},
+                {"override_body": "a {{runs}}"},
                 "test-message",
                 "warning",
                 {
-                    "body": "params_message",
-                    "subject": "[warning] params_message",
+                    "subject": "[warning] test-message",
+                    "body": MOCKED_BODY_PLACEHOLDER,
                 },
             ),
         ],
@@ -1212,6 +1214,10 @@ class TestMailNotification:
         logger.debug(f"Testing {name}")
         notification = mail.MailNotification(params=params)
         notification._send_email = unittest.mock.AsyncMock()
+        notification._get_html = unittest.mock.MagicMock(return_value=self.MOCKED_HTML)
+        notification._serialize_runs_in_request_body = unittest.mock.MagicMock(
+            return_value=self.MOCKED_BODY_PLACEHOLDER
+        )
         await notification.push(message, severity, [])
         assert notification.params["subject"] == expected["subject"]
         assert notification.params["body"] == expected["body"]
